+++ lib/project_types/theme/commands/package.rb
@@ -18,11 +18,15 @@ module Theme
         release-notes.md
       ]
 
+      ZIP = "zip"
+      SEVEN_ZIP = "7z"
+
       def call(args, _name)
         path = args.first || "."
 
-        check_prereq_command("zip")
+        check_prereq_command
         zip_name = theme_name(path) + ".zip"
+
         zip(zip_name, path, THEME_DIRECTORIES)
         @ctx.done(@ctx.message("theme.package.done", zip_name))
       end
@@ -33,13 +37,12 @@ module Theme
 
       private
 
-      def check_prereq_command(command)
-        cmd_path = @ctx.which(command)
-        @ctx.abort(@ctx.message("theme.package.error.prereq_command_required", command)) if cmd_path.nil?
+      def check_prereq_command
+        @ctx.abort(@ctx.message("theme.package.error.prereq_command_required")) if command.nil?
       end
 
       def zip(zip_name, path, files)
-        @ctx.system("zip", "-r", zip_name, *files, chdir: path)
+        @ctx.system(command, command_flags, zip_name, *files, chdir: path)
       end
 
       def theme_name(path)
@@ -52,6 +55,18 @@ module Theme
         @ctx.abort(@ctx.message("theme.package.error.missing_theme_name")) unless theme_name
 
         [theme_name, theme_info["theme_version"]].compact.join("-")
+      end
+
+      def command
+        @command ||= if @ctx.which(ZIP)
+          ZIP
+        elsif @ctx.which(SEVEN_ZIP)
+          SEVEN_ZIP
+        end
+      end
+
+      def command_flags
+        @command_flags ||= command == ZIP ? "-r" : "a"
       end
     end
   end